feat: Use Prettier and git blame plugins, disable clipboard sync
Work requires that I use Prettier for formatting on save (the Git
conflicts caused by someone not having Prettier are said to be a
pain). So much for relying on tsserver instead of a global Prettier
install and a plugin. This commit also adds another work must-have,
a blame plugin. Finally, I kind of like keeping the system and
Vim keyboards separate, so that setting is disabled.